AG-12600 Add disjoint channel drawing text label
